+++ src/helpers/__tests__/scrollable.js
@@ -1,0 +1,8 @@
+import scrollable from "../scrollable";
+
+it("allows for scrolling on web", () => {
+  expect(scrollable.resolve()).toEqual({
+    overflowY: "scroll",
+    WebkitOverflowScrolling: "touch",
+  });
+});